fix: Drop the unreachable settings-URL fallback and share the URL recorder
## Summary
- The unanchored Privacy & Security URL could never succeed where the
  anchored one failed: `NSWorkspace.open` reports that Launch Services
  resolved and launched a handler for the *scheme*, and both URLs carry
  the same scheme and pane id. A dropped `?Privacy_Microphone` anchor
  still opens System Settings and still returns `true`, so the fallback
  was dead for exactly the case it was written for. Replaced with the
  return-value contract stated on the symbol.

## Changes
- `SystemSettingsLink`: one URL constant, and a `///` recording that the
  return value means "System Settings launched", not "reached the
  Microphone pane" — an unhonored anchor degrades to the pane root, which
  the popover's written steps already cover.
- `URLOpenRecorder` moved from `SystemSettingsLinkTests.swift` to
  `KernovaTests/Mocks/`; three suites share it.
- Dropped the two tests asserting the fallback ordering.

## Test plan
- [x] `make build`
- [x] `make test` — 2834 cases pass
- [x] `make lint`
